+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added events endpoint to retrieve events that are occurring in your Smartsheet plan.
 
 ## [1.3.0] - 2019-02-11
 - Added constant for smartsheetgov